% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorite_artists.R
\name{favorite_artists}
\alias{favorite_artists}
\title{Count the number of artists in your playlist}
\usage{
favorite_artists(spotify_data, count = 6L, see_all = FALSE, ascending = FALSE)
}
\arguments{
\item{spotify_data}{A data.frame from exportify.net}

\item{count}{An integer indicating the number of rows to return}

\item{see_all}{A Boolean indicating if the function should return all available rows}

\item{ascending}{A Boolean who orders the resulting data.frame based on num_appearances}
}
\value{
A data.frame ordering the amount of times each artist appears in a data.frame
}
\description{
Count the number of artists in your playlist
}
\examples{
data('rb')
favorite_artists(rb, see_all = TRUE)

}
