% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorite_artists.R
\name{favorite_artists}
\alias{favorite_artists}
\title{Count the number of artists in your playlist!}
\usage{
favorite_artists(spotify_data, count = 6L, see_all = FALSE, ascending = FALSE)
}
\arguments{
\item{spotify_data}{A data.frame from exportify.net}

\item{count}{An integer who dictates the number of rows}

\item{see_all}{A Boolean who dictates if the function should return all available rows}
}
\value{
A data.frame ordering the amount of times each artist appears in a data.frame
}
\description{
Count the number of artists in your playlist!
}
\examples{
my_playlist <- readr::read_csv('../randb.csv')
favorite_artists(spotify_data = my_playlist)
favorite_artists(my_playlist, count = 5L)

}
